corr

+++ src/okulardocument.cpp
@@ -99,7 +99,7 @@ void OkularDocument::adjustSize(int &width, int &height)
   if((long)width*(long)height > 20000000L) {
     qDebug() << "adjust width and height in order to stay below threshold";
     //adjust height in order to stay below this threshold (used by okular core library)
-    qreal factor2 = 20000000L/((long)width*(long)height);
+    qreal factor2 = qreal(20000000L)/qreal((long)width*(long)height);
     height = int(height*factor2);
   }
 }
